Extract the pure tiling layers into a fibsem.imaging.tiling package (FIB-390)
tiled.py mixed three concerns: grid geometry, stage/image reprojection, and
matplotlib rendering -- none of which need a microscope -- with the acquisition
runner, which does. The fluorescence tiler needs the first three and has its own
copy of the geometry, which is how the two drifted into opposite row-stepping
conventions and produced a reversed mosaic (#226).

Split into fibsem/imaging/tiling/:

  geometry.py      TilePosition, compute_tile_grid, order_tiles, _spiral_order,
                   validate_tile_stage_positions
  reprojection.py  both reprojection variants, the coordinate-system helpers,
                   _inverse_y_corrected_stage_movement
  plotting.py      plot_tile_positions, plot_stage_positions_on_image,
                   plot_minimap, POSITION_COLOURS

Plotting is separate so that consuming tile geometry does not cost a matplotlib
import. There was already a "TODO: move these to fibsem.imaging.reprojection?"
above the reprojection block; this does that.

Staying in tiled.py: TiledAcquisitionRunner, its cancellation helper, the
tiled_image_acquisition_and_stitch wrapper, and convert_image_coord_to_stage_position
and its plural. Those last two look pure but are not -- they take a microscope and
project through project_stable_move. Note the asymmetry: stage-to-image is pure
because it uses the baked-in inverse, image-to-stage needs the live projection.

tiled.py re-exports everything moved, so its 15 importers, the star-import in
fibsem/imaging/__init__.py, and the visual harness in fibsem/imaging/tests/ are
unaffected.

Every moved definition was verified AST-identical to its original rather than
retyped, and the reprojection functions -- which had no direct coverage -- were
characterised first, so the move is demonstrably inert.

On the motivation: this was originally justified partly by an import cycle said to
prevent fibsem.fm from importing fibsem.imaging.tiled. That cycle does not exist.
fm.acquisition can import tiled today, unmodified, because the back-edge
(fibsem.microscope -> fibsem.fm.microscope) is a submodule import that never needs
the partially-initialised package's attributes. The reason to share one definition
is the one that always applied: two copies of the same geometry had already drifted
into a shipped bug.

The independence is enforced by a static check on declared imports rather than at
runtime. A runtime check cannot separate a module's own dependencies from its
parent package's -- fibsem/imaging/__init__.py star-imports tiled, so importing any
submodule of fibsem.imaging loads the microscope no matter what that submodule
needs. Removing that star-import is a public API change and does not belong here.
